Fix undefined behavior in libgifcodec
SkGifImageReader::parse has a single parameter that is used either to
represent a frame index or an SkGIFParseQuery value. It is undefined to
cast an out of range integer as an SkGIFParseQuery, so change the
parameter to an int, which can be cast to an SkGIFParseQuery.

Change-Id: I4b321f117cd8bc969686f3c1c6bf35768644f193
Reviewed-on: https://skia-review.googlesource.com/c/libgifcodec/+/334841
Auto-Submit: Leon Scroggins <scroggo@google.com>
Reviewed-by: Julia Lavrova <jlavrova@google.com>
Commit-Queue: Leon Scroggins <scroggo@google.com>