Ensure frame rect is on screen or empty
Bug: b/160984428

There is no reason for a client to need to know that the frame rect
extends beyond the edge of the "canvas"/"screen"/image bounds. A typical
use case is more interested in knowing what part of the original bounds
the frame update occupies.

If the offsets are beyond the image bounds, move them to the
right/bottom edge of the image (and make the width/height 0). This
matches Wuffs.

Change-Id: I215b312b421a30bf985f681424a4c4198f069862
Reviewed-on: https://skia-review.googlesource.com/c/libgifcodec/+/339831
Commit-Queue: Leon Scroggins <scroggo@google.com>
Commit-Queue: Derek Sollenberger <djsollen@google.com>
Auto-Submit: Leon Scroggins <scroggo@google.com>
Reviewed-by: Derek Sollenberger <djsollen@google.com>